refactor: 인수 테스트 과정에서 불필요한 DB 초기화 쿼리 실행 개선 (#1878)
* fix: AUTO_INCREMENT 초기화 조건 수정

- 1 이상 -> 1 초과

* test: 테이블 초기화 중복 로직 삭제

- forceVerify 이후에는 setup 메소드만 선언
- 이외의 clear와 setup이 같이 선언되어 있으면 삭제

* test: 테이블 Truncate 메소드 변경

- 모든 테이블 Truncate -> AUTO_INCREMENT가 1 초과 혹은 NULL인 테이블 Truncate

* chore: 미사용 필드 삭제

* test: TRUNCATE 쿼리 조건 변경

- AUTO_INCREMENT 조건이 아닌 COUNT(*) 조건으로 변경

* chore: 메소드 위치 수정

+++ src/test/java/in/koreatech/koin/acceptance/support/DBInitializer.java
@@ -3,8 +3,6 @@ package in.koreatech.koin.acceptance.support;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.sql.DataSource;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.TestComponent;
 import org.springframework.data.mongodb.core.MongoTemplate;
@@ -20,12 +18,6 @@ public class DBInitializer {
 
     private static final int OFF = 0;
     private static final int ON = 1;
-    private static final int COLUMN_INDEX = 1;
-
-    private List<String> tableNames = new ArrayList<>();
-
-    @Autowired
-    private DataSource dataSource;
 
     @PersistenceContext
     private EntityManager entityManager;
@@ -36,25 +28,34 @@ public class DBInitializer {
     @Autowired
     private MongoTemplate mongoTemplate;
 
+    private List<String> tableNames = new ArrayList<>();
+
     private void findDatabaseTableNames() {
         String sql = "SHOW TABLES";
         tableNames = entityManager.createNativeQuery(sql).getResultList();
     }
 
-    private void truncateAllTable() {
+    private void truncateDirtyTables() {
         setForeignKeyCheck(OFF);
-        for (String tableName: tableNames) {
-            entityManager.createNativeQuery(String.format("TRUNCATE TABLE `%s`", tableName)).executeUpdate();
+        for (String tableName : tableNames) {
+            long count = ((Number) entityManager
+                .createNativeQuery(String.format("SELECT COUNT(*) FROM `%s`", tableName))
+                .getSingleResult()).longValue();
+
+            if (count > 0) {
+                entityManager.createNativeQuery(String.format("TRUNCATE TABLE `%s`", tableName)).executeUpdate();
+            }
         }
         setForeignKeyCheck(ON);
     }
 
     @Transactional
     public void initIncrement() {
-        String sql = "SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = 'test' AND AUTO_INCREMENT >= 1";
+        String sql = "SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = 'test' AND AUTO_INCREMENT > 1";
         List<String> dirtyTables = entityManager.createNativeQuery(sql).getResultList();
-        for (String tableName: dirtyTables) {
-            entityManager.createNativeQuery(String.format("ALTER TABLE %s AUTO_INCREMENT = 1", tableName)).executeUpdate();
+        for (String tableName : dirtyTables) {
+            entityManager.createNativeQuery(String.format("ALTER TABLE %s AUTO_INCREMENT = 1", tableName))
+                .executeUpdate();
         }
     }
 
@@ -68,7 +69,7 @@ public class DBInitializer {
             findDatabaseTableNames();
         }
         entityManager.clear();
-        truncateAllTable();
+        truncateDirtyTables();
         clearRedis();
         clearMongo();
     }
